added monospace font CourierPrimeSans and split main.rs into style.rs + app.rs
