+++ src/main.rs
@@ -4,6 +4,8 @@ mod info_address_port_pair;
 mod args;
 mod thread_write_report_functions;
 mod info_traffic;
+mod style;
+mod app;
 
 use std::fs::File;
 use std::io::prelude::*;
@@ -32,7 +34,7 @@ use iced::futures::FutureExt;
 use iced_style::pane_grid::Line;
 use indexmap::IndexMap;
 use crate::info_traffic::InfoTraffic;
-use crate::style::Mode;
+use style::{Mode, FontSizeBody, FontSizeSubtitle, FontSizeTitle, icon_sun_moon};
 
 
 pub struct Filters {
@@ -110,591 +112,29 @@ pub fn main() {
                            mutex_map1, status_pair1);
     });
 
-    Sniffer::run(Settings::with_flags(Sniffer {
-        info_traffic: mutex_map3,
-        device: found_device3,
-        filters: filters3,
-        status_pair: status_pair3,
-        start: button::State::new(),
-        reset: button::State::new(),
-        mode: button::State::new(),
-        report: button::State::new(),
-        app: pick_list::State::new(),
-        scroll: scrollable::State::new(),
-        style: Mode::Night
-    }));
+    Sniffer::run(Settings {
+        id: None,
+        window: Default::default(),
+        flags: Sniffer {
+            info_traffic: mutex_map3,
+            device: found_device3,
+            filters: filters3,
+            status_pair: status_pair3,
+            start: button::State::new(),
+            reset: button::State::new(),
+            mode: button::State::new(),
+            report: button::State::new(),
+            app: pick_list::State::new(),
+            scroll: scrollable::State::new(),
+            style: Mode::Night
+        },
+        default_font: Some(include_bytes!("../fonts/CourierPrimeSans.ttf")),
+        default_text_size: FontSizeBody,
+        text_multithreading: false,
+        antialiasing: false,
 
-}
+        exit_on_close_request: true,
+        try_opengles_first: false
+    });
 
-#[derive(Debug, Clone)]
-enum Message {
-    Tick,
-    AdapterSelection(String),
-    IpVersionSelection(String),
-    TransportProtocolSelection(TransProtocol),
-    AppProtocolSelection(AppProtocol),
-    OpenReport,
-    Start,
-    Reset,
-    Style
-}
-
-impl Application for Sniffer {
-    type Executor = executor::Default;
-    type Message = Message;
-    type Flags = Sniffer;
-
-    fn new(flags: Sniffer) -> (Sniffer, Command<Message>) {
-        (
-            flags,
-            Command::none(),
-        )
-    }
-
-    fn title(&self) -> String {
-        String::from("Sniffnet")
-    }
-
-    fn update(&mut self, message: Message) -> Command<Message> {
-        match message {
-            Message::Tick => {}
-            Message::AdapterSelection(name) => {
-                *self.device.lock().unwrap() = Device::from(&*name);
-            }
-            Message::IpVersionSelection(version) => {
-                self.filters.lock().unwrap().ip = version;
-            }
-            Message::TransportProtocolSelection(protocol) => {
-                self.filters.lock().unwrap().transport = protocol;
-            }
-            Message::AppProtocolSelection(protocol) => {
-                self.filters.lock().unwrap().application = protocol;
-            }
-            Message::OpenReport => {
-                #[cfg(target_os = "windows")]
-                let command = "explorer";
-                #[cfg(target_os = "macos")]
-                let command = "open";
-                #[cfg(target_os = "linux")]
-                let command = "explorer";
-                std::process::Command::new( command )
-                    .arg( "./sniffnet_report/report.txt" )
-                    .spawn( )
-                    .unwrap( );
-            }
-            Message::Start => {
-                *self.status_pair.0.lock().unwrap() = Status::Running;
-                &self.status_pair.1.notify_all();
-            }
-            Message::Reset => {
-                let mut info_traffic = self.info_traffic.lock().unwrap();
-                *info_traffic = InfoTraffic::new();
-                info_traffic.reset();
-                *self.status_pair.0.lock().unwrap() = Status::Init;
-            }
-            Message::Style => {
-                self.style = if self.style == Mode::Day {
-                    Mode::Night
-                }
-                else {
-                    Mode::Day
-                };
-            }
-        }
-
-        Command::none()
-    }
-
-    fn subscription(&self) -> Subscription<Message> {
-        match self.status_pair.0.lock().unwrap() {
-            _ => {
-                iced::time::every(Duration::from_millis(1000)).map(|_| Message::Tick)
-            }
-        }
-    }
-
-    fn view(&mut self) -> Element<Message> {
-
-        let button_start = Button::new(
-            &mut self.start,
-            Text::new("Run!").vertical_alignment(alignment::Vertical::Center).horizontal_alignment(alignment::Horizontal::Center),
-            )
-            .padding(10)
-            .height(Length::Units(80))
-            .width(Length::Units(160))
-            .style(self.style)
-            .on_press(Message::Start);
-
-        let button_reset = Button::new(
-            &mut self.reset,
-            Text::new("Reset").horizontal_alignment(alignment::Horizontal::Center),
-            )
-            .padding(10)
-            .height(Length::Units(40))
-            .width(Length::Units(80))
-            .style(self.style)
-            .on_press(Message::Reset);
-
-        let button_style = Button::new(
-            &mut self.mode,
-            // icon(match self.style {
-            //     Mode::Night => {font_awesome::SUN}
-            //     Mode::Day => {font_awesome::MOON}
-                icon_sun_moon(self.style)
-            .horizontal_alignment(alignment::Horizontal::Center),
-            )
-            .padding(10)
-            .height(Length::Units(40))
-            .width(Length::Units(100))
-            .style(self.style)
-            .on_press(Message::Style);
-
-        let button_report = Button::new(
-            &mut self.report,
-            Text::new("Open network traffic report").horizontal_alignment(alignment::Horizontal::Center),
-        )
-            .padding(10)
-            .height(Length::Units(40))
-            .width(Length::Units(270))
-            .style(self.style)
-            .on_press(Message::OpenReport);
-
-        let logo = Svg::from_path("./resources/sniffnet_logo.svg", );
-
-        let header = Row::new()
-            .height(Length::FillPortion(3))
-            .align_items(Alignment::Center)
-            .push(Column::new().width(Length::FillPortion(1)))
-            .push(Column::new().width(Length::FillPortion(6)).push(logo))
-            .push(Row::new().width(Length::FillPortion(1)).align_items(Alignment::Center).push(button_style));
-
-        let mut dev_str_list = vec![];
-        for dev in Device::list().expect("Error retrieving device list\r\n") {
-            let mut dev_str = String::new();
-            match dev.desc {
-                None => {
-                    dev_str.push_str(&format!("Device:  {}", dev.name));
-                }
-                Some(description) => {
-                    dev_str.push_str(&format!("Device:  {}\nDescription:  {}", dev.name.cyan(), description));
-                }
-            }
-            let num_addresses = dev.addresses.len();
-            match num_addresses {
-                0 => {},
-                1 => {dev_str.push_str("\nAddress:  ");},
-                _ => {dev_str.push_str("\nAddresses:  ");}
-            }
-
-            let mut x = 0;
-            for addr in dev.addresses {
-                x += 1;
-                let address_string = addr.addr.to_string();
-                if x == num_addresses {
-                    dev_str.push_str(&format!("{}", address_string));
-                }
-                else {
-                    dev_str.push_str(&format!("{}\n                           ", address_string));
-                }
-            }
-            dev_str_list.push((dev.name, dev_str));
-        }
-
-        let col_adapter = Column::new()
-            .padding(20)
-            .spacing(10)
-            .height(Length::Fill)
-            .width(Length::FillPortion(4))
-            .push(Text::new("Select network adapter to inspect").size(32))
-            .push(dev_str_list.iter().fold(
-                Scrollable::new(&mut self.scroll).style(self.style).padding(10).spacing(20).height(Length::FillPortion(8)),
-                |scroll, adapter| {
-                    scroll.push(Radio::new(
-                        &adapter.0,
-                        &adapter.1,
-                        Some(&self.device.clone().lock().unwrap().name),
-                        |name| Message::AdapterSelection(name.to_string()),
-                    ).size(15).style(self.style))
-                },
-            ));
-
-        let col_space = Column::new()
-            .padding(20)
-            .spacing(10)
-            .width(Length::FillPortion(1));
-
-        let filtri = self.filters.lock().unwrap();
-        let ip_active = &*filtri.ip;
-        let col_ip_radio = Column::new().spacing(10)
-            .push(Text::new("IP version").size(24))
-            .push(Radio::new(
-                "ipv4",
-                "IPv4",
-                Some(ip_active),
-                |version| Message::IpVersionSelection(version.to_string())
-            ).size(15).style(self.style))
-            .push(Radio::new(
-                "ipv6",
-                "IPv6",
-                Some(ip_active),
-                |version| Message::IpVersionSelection(version.to_string())
-            ).size(15).style(self.style))
-            .push(Radio::new(
-                "no filter",
-                "both",
-                Some(ip_active),
-                |version| Message::IpVersionSelection(version.to_string())
-            ).size(15).style(self.style));
-        let col_ip = Column::new()
-            .spacing(10)
-            .width(Length::FillPortion(1))
-            .push(col_ip_radio);
-
-        let transport_active = filtri.transport;
-        let col_transport_radio = Column::new().spacing(10)
-            .push(Text::new("Transport protocol").size(24))
-            .push(Radio::new(
-                TransProtocol::TCP,
-                "TCP",
-                Some(transport_active),
-                |protocol| Message::TransportProtocolSelection(protocol)
-            ).size(15).style(self.style))
-            .push(Radio::new(
-                TransProtocol::UDP,
-                "UDP",
-                Some(transport_active),
-                |protocol| Message::TransportProtocolSelection(protocol)
-            ).size(15).style(self.style))
-            .push(Radio::new(
-                TransProtocol::Other,
-                "both",
-                Some(transport_active),
-                |protocol| Message::TransportProtocolSelection(protocol)
-            ).size(15).style(self.style));
-        let col_transport = Column::new()
-            .align_items(Alignment::Center)
-            .spacing(10)
-            .width(Length::FillPortion(2))
-            .push(col_transport_radio)
-            .push(Row::new().height(Length::FillPortion(2)))
-            .push(button_start)
-            .push(Row::new().height(Length::FillPortion(1)));
-
-        let app_active = filtri.application;
-        let picklist_app = PickList::new(
-            &mut self.app,
-            &AppProtocol::ALL[..],
-            Some(app_active),
-            |protocol| Message::AppProtocolSelection(protocol),
-        )
-            .placeholder("Select application protocol")
-            .style(self.style);
-        let mut col_app = Column::new()
-            .width(Length::FillPortion(2))
-            .spacing(10)
-            .push(iced::Text::new("Application protocol").size(24))
-            .push(picklist_app);
-
-        let filters = Column::new().width(Length::FillPortion(6)).padding(20).spacing(20)
-            .push(Row::new().push(Text::new("Select filters to be applied on network traffic").size(32)))
-            .push(Row::new().height(Length::FillPortion(3)).push(col_ip).push(col_transport).push(col_app));
-
-
-        let sniffer = self.info_traffic.lock().unwrap();
-
-        let mut col_packets = Column::new()
-            .width(Length::FillPortion(1))
-            .align_items(Alignment::Center)
-            .spacing(20)
-            .push(iced::Text::new(std::env::current_dir().unwrap().to_str().unwrap()))
-            .push(iced::Text::new(sniffer.all_packets.to_string()));
-        if sniffer.tot_received_packets + sniffer.tot_sent_packets > 0 {
-            col_packets = col_packets
-                .push(iced::Text::new("Packets count per application protocol"))
-                .push(iced::Text::new(get_app_count_string(sniffer.app_protocols.clone(), sniffer.tot_received_packets + sniffer.tot_sent_packets)));
-        }
-        col_packets = col_packets.push(button_reset).push(button_report);
-
-        let mut body = Row::new().height(Length::FillPortion(9));
-
-        match *self.status_pair.0.lock().unwrap() {
-            Status::Init => {body = body
-                .push(col_adapter)
-                .push(col_space)
-                .push(filters);}
-            Status::Running => {body = body
-                .push(col_packets);}
-            Status::Pause => {}
-            Status::Stop => {}
-        }
-
-        Container::new(Column::new().push(header).push(body))
-            .width(Length::Fill)
-            .height(Length::Fill)
-            .center_x()
-            .center_y()
-            .style(self.style)
-            .into()
-    }
-
-}
-
-mod style {
-    use iced::{pick_list, container, Background, Color, Vector, Container, Element, Row, Application, button};
-    use iced::container::{Style, StyleSheet};
-    use iced_style::scrollable::{Scrollbar, Scroller};
-    use crate::Message;
-
-    #[derive(Copy, Eq, PartialEq)]
-    pub enum Mode {
-        Night,
-        Day
-    }
-
-    impl Clone for Mode {
-        fn clone(&self) -> Self {
-            *self
-        }
-    }
-
-    impl StyleSheet for Mode {
-        fn style(&self) -> Style {
-            Style {
-                text_color: match self {
-                    Mode::Day => Some(Color::BLACK),
-                    Mode::Night => Some(Color::WHITE),
-                },
-                background: match self {
-                    Mode::Day => {Some(Background::Color(Color{r: 0.8, g: 0.8, b: 0.8, a: 1.0,}))}
-                    Mode::Night => {Some(Background::Color(Color{r: 0.2, g: 0.2, b: 0.2, a: 1.0,}))}
-                },
-                border_radius: 0.0,
-                border_width: 0.0,
-                border_color: Default::default()
-            }
-        }
-    }
-
-    impl pick_list::StyleSheet for Mode {
-        fn menu(&self) -> iced_style::menu::Style {
-            iced_style::menu::Style {
-                text_color: match self {
-                    Mode::Day => Color::BLACK,
-                    Mode::Night => Color::WHITE,
-                },
-                background: Background::Color(match self {
-                Mode::Day => Color{r: 0.9, g: 0.9, b: 0.9, a: 1.0,},
-                Mode::Night => Color{r: 0.1, g: 0.1, b: 0.1, a: 1.0,},
-                }),
-                border_width: 2.0,
-                border_color: match self {
-                    Mode::Day => Color{r: 0.0, g: 0.5, b: 0.8, a: 1.0,},
-                    Mode::Night => Color{r: 0.0, g: 0.8, b: 0.5, a: 1.0,},
-                },
-                selected_text_color: match self {
-                    Mode::Day => Color::BLACK,
-                    Mode::Night => Color::WHITE,
-                },
-                selected_background: Background::Color(match self {
-                    Mode::Day => Color{r: 0.8, g: 0.8, b: 0.8, a: 1.0,},
-                    Mode::Night => Color{r: 0.2, g: 0.2, b: 0.2, a: 1.0,},
-                })
-            }
-        }
-
-        fn active(&self) -> pick_list::Style {
-            pick_list::Style {
-                text_color: match self {
-                    Mode::Day => Color::BLACK,
-                    Mode::Night => Color::WHITE,
-                },
-                placeholder_color: Color::BLACK,
-                background: Background::Color(match self {
-                    Mode::Day => Color{r: 0.9, g: 0.9, b: 0.9, a: 1.0,},
-                    Mode::Night => Color{r: 0.1, g: 0.1, b: 0.1, a: 1.0,},
-                }),
-                border_radius: 0.0,
-                border_width: 2.0,
-                border_color: match self {
-                    Mode::Day => Color{r: 0.0, g: 0.5, b: 0.8, a: 1.0,},
-                    Mode::Night => Color{r: 0.0, g: 0.8, b: 0.5, a: 1.0,},
-                },
-                icon_size: 0.5
-            }
-        }
-
-        fn hovered(&self) -> pick_list::Style {
-            pick_list::Style {
-                text_color: match self {
-                    Mode::Day => Color::BLACK,
-                    Mode::Night => Color::WHITE,
-                },
-                placeholder_color: Color::BLACK,
-                background: Background::Color(match self {
-                    Mode::Day => Color{r: 0.8, g: 0.8, b: 0.8, a: 1.0,},
-                    Mode::Night => Color{r: 0.2, g: 0.2, b: 0.2, a: 1.0,},
-                }),
-                border_radius: 0.0,
-                border_width: 2.0,
-                border_color: match self {
-                    Mode::Day => Color{r: 0.0, g: 0.5, b: 0.5, a: 1.0,},
-                    Mode::Night => Color{r: 0.0, g: 0.5, b: 0.5, a: 1.0,},
-                },
-                icon_size: 0.5
-            }
-        }
-    }
-
-    impl button::StyleSheet for Mode {
-
-        fn hovered(&self) -> iced_style::button::Style {
-            iced_style::button::Style {
-                shadow_offset: Vector::new(1.0, 1.0),
-                background: Some(Background::Color(match self {
-                    Mode::Day => Color{r: 0.8, g: 0.8, b: 0.8, a: 1.0,},
-                    Mode::Night => Color{r: 0.2, g: 0.2, b: 0.2, a: 1.0,},
-                })),
-                border_radius: 12.0,
-                border_width: 2.0,
-                border_color: match self {
-                    Mode::Day => Color{r: 0.0, g: 0.5, b: 0.5, a: 1.0,},
-                    Mode::Night => Color{r: 0.0, g: 0.5, b: 0.5, a: 1.0,},
-                },
-                text_color: match self {
-                    Mode::Day => Color::BLACK,
-                    Mode::Night => Color::WHITE,
-                }
-            }
-        }
-
-        fn active(&self) -> button::Style {
-            button::Style {
-                background: Some(Background::Color(match self {
-                    Mode::Day => Color{r: 0.9, g: 0.9, b: 0.9, a: 1.0,},
-                    Mode::Night => Color{r: 0.1, g: 0.1, b: 0.1, a: 1.0,},
-                })),
-                border_radius: 12.0,
-                border_width: 2.0,
-                shadow_offset: Vector::new(0.0, 0.0),
-                text_color:  match self {
-                    Mode::Day => Color::BLACK,
-                    Mode::Night => Color::WHITE,
-                },
-                border_color: match self {
-                    Mode::Day => Color{r: 0.0, g: 0.5, b: 0.8, a: 1.0,},
-                    Mode::Night => Color{r: 0.0, g: 0.8, b: 0.5, a: 1.0,},
-                }
-            }
-        }
-    }
-
-    impl iced_style::radio::StyleSheet for Mode {
-        fn active(&self) -> iced_style::radio::Style {
-            iced_style::radio::Style {
-                background: Background::Color(match self {
-                    Mode::Day => Color{r: 1.0, g: 1.0, b: 1.0, a: 1.0,},
-                    Mode::Night => Color{r: 0.1, g: 0.1, b: 0.1, a: 1.0,},
-                }),
-                dot_color: match self {
-                    Mode::Day => Color{r: 0.0, g: 0.5, b: 0.8, a: 1.0,},
-                    Mode::Night => Color{r: 0.0, g: 0.8, b: 0.5, a: 1.0,},
-                },
-                border_width: 0.0,
-                border_color: Default::default(),
-                text_color: None
-            }
-        }
-
-        fn hovered(&self) -> iced_style::radio::Style {
-            iced_style::radio::Style {
-                background: Background::Color(match self {
-                    Mode::Day => Color{r: 1.0, g: 1.0, b: 1.0, a: 1.0,},
-                    Mode::Night => Color{r: 0.1, g: 0.1, b: 0.1, a: 1.0,},
-                }),
-                dot_color: match self {
-                    Mode::Day => Color{r: 0.0, g: 0.5, b: 0.8, a: 1.0,},
-                    Mode::Night => Color{r: 0.0, g: 0.8, b: 0.5, a: 1.0,},
-                },
-                border_width: 2.0,
-                border_color: match self {
-                    Mode::Day => Color{r: 0.0, g: 0.5, b: 0.8, a: 1.0,},
-                    Mode::Night => Color{r: 0.0, g: 0.8, b: 0.5, a: 1.0,},
-                },
-                text_color: None
-            }
-        }
-    }
-
-    impl iced_style::scrollable::StyleSheet for Mode {
-        fn active(&self) -> Scrollbar {
-            Scrollbar {
-                background: Some(Background::Color(match self {
-                    Mode::Day => Color{r: 0.9, g: 0.9, b: 0.9, a: 1.0,},
-                    Mode::Night => Color{r: 0.1, g: 0.1, b: 0.1, a: 1.0,},
-                })),
-                border_radius: 12.0,
-                border_width: 0.0,
-                border_color: Color::BLACK,
-                scroller: Scroller {
-                    color: match self {
-                        Mode::Day => Color{r: 0.8, g: 0.8, b: 0.8, a: 1.0,},
-                        Mode::Night => Color{r: 0.2, g: 0.2, b: 0.2, a: 1.0,},
-                    },
-                    border_radius: 12.0,
-                    border_width: 2.0,
-                    border_color: Color::BLACK
-                }
-            }
-        }
-
-        fn hovered(&self) -> Scrollbar {
-            Scrollbar {
-                background: Some(Background::Color(match self {
-                    Mode::Day => Color{r: 0.9, g: 0.9, b: 0.9, a: 1.0,},
-                    Mode::Night => Color{r: 0.1, g: 0.1, b: 0.1, a: 1.0,},
-                })),
-                border_radius: 12.0,
-                border_width: 2.0,
-                border_color: Color::BLACK,
-                scroller: Scroller {
-                    color: match self {
-                        Mode::Day => Color{r: 0.0, g: 0.5, b: 0.8, a: 1.0,},
-                        Mode::Night => Color{r: 0.0, g: 0.8, b: 0.5, a: 1.0,},
-                    },
-                    border_radius: 12.0,
-                    border_width: 2.0,
-                    border_color: Color::BLACK
-                }
-            }
-        }
-    }
-
-}
-
-const ICONS: Font = Font::External {
-    name: "Icons",
-    bytes: include_bytes!("../fonts/icons.ttf"),
-};
-
-fn icon(unicode: char) -> Text {
-    Text::new(unicode.to_string())
-        .font(ICONS)
-        .width(Length::Units(20))
-        .horizontal_alignment(alignment::Horizontal::Center)
-        .size(20)
-}
-
-fn icon_sun_moon(style: Mode) -> Text {
-    match style {
-        Mode::Night => {    Text::new(format!("{} {} {}", font_awesome::MOON, font_awesome::ANGLE_DOUBLE_RIGHT, font_awesome::SUN))
-            .font(ICONS)
-            .width(Length::Units(20))
-            .horizontal_alignment(alignment::Horizontal::Center)
-            .size(20)}
-        Mode::Day => {    Text::new(format!("{} {} {}", font_awesome::SUN, font_awesome::ANGLE_DOUBLE_RIGHT, font_awesome::MOON))
-            .font(ICONS)
-            .width(Length::Units(20))
-            .horizontal_alignment(alignment::Horizontal::Center)
-            .size(20)}
-    }
 }
